scripts: Added steps for removing build directory and *.egg-info directory in ci-test.sh script.

+++ scripts/ci-test.sh
@@ -2,8 +2,17 @@
 
 set -e
 
-rm -rf build/
-pip uninstall -y pypinball
+if [ -d build/ ]; then
+  echo "Removing build/ directory"
+  rm -rf build/
+fi
+
+if [ -d *.egg-info ]; then
+  echo "Removing any/all *.egg-info directories"
+  rm -rf *.egg-info
+fi
+
+pip uninstall -y -q pypinball
 ./scripts/black-formatter.sh
 ./scripts/run-tests.sh
 ./scripts/build-docs.sh
